+++ R/autoItemValuesMinMax.R
@@ -2,12 +2,14 @@
 #'
 #' \code{\link{itemValuesDeviation}} creates constraints related to an item parameter/value. \code{autoItemValuesMixMax} automatically
 #' determines the appropriate \code{targetValue} and then calls \code{\link{itemValuesDeviation}}. The function only works for
-#' (dichotomous) dummy indicators with values 0 or 1 and if the item pool is fully depleted in the test assembly.
+#' (dichotomous) dummy indicators with values 0 or 1.
 #'
-#' Two scenarios are possible when automatically determining the target value: (a) Either items with the selected property could be exactly
+#' Two scenarios are possible when automatically determining the target value:
+#' (a) Either items with the selected property could be exactly
 #' distributed across test forms or (b) this is not possible. An example would be 2 test forms and 4 multiple choice items (a) or 2 test
 #' forms and 5 multiple choice items (b). If (a), the tolerance level works exactly as one would expect. If (b) the tolerance level is
-#' adapted, meaning that if tolerance level is 0 in example (b), allowed values are 2 or 3 multiple choice items per test form.
+#' adapted, meaning that if tolerance level is 0 in example (b), allowed values are 2 or 3 multiple choice items per test form. For detailed documentation on how the minimum and maximum are calculated
+#' see also \code{\link{computeTargetValues}}.
 #'
 #' @inheritParams itemValuesConstraint
 #' @inheritParams computeTargetValues
@@ -19,11 +21,11 @@
 #' autoItemValuesMinMax(2, itemValues = c(0, 1, 0, 1))
 #'
 #' @export
-autoItemValuesMinMax <- function(nForms, itemValues, allowedDeviation = NULL,
+autoItemValuesMinMax <- function(nForms, itemValues, testLength = NULL, allowedDeviation = NULL,
                                  relative = FALSE, verbose = TRUE, itemIDs = NULL){
   check_itemIDs(itemIDs)
   # compute the minimum and maximum values
-  min_max <- computeTargetValues(itemValues, nForms, testLength = NULL,
+  min_max <- computeTargetValues(itemValues, nForms, testLength = testLength,
                                  allowedDeviation = allowedDeviation, relative = relative)
   min_max <- round(min_max, 2)
   if(min_max[1] < 0) min_max[1] <- 0