+++ NEWS.md
@@ -4,6 +4,7 @@
 * `itemTargetConstraint()` substituted by `minimaxConstraint()`, `maximinConstraint()`, `maxConstraint()`, `minConstraint()`, `cappedMaximinConstraint()` which allow for different types of optimization goals
 * `Symphony` added as available solver (via `Rsymphony`)
 * `inspectSolution()` now allows character variables in the input if sums are calculated
+* `autoItemValuesMinMax()` now works for different test length via `testLength` argument
 
 ## Documentation